services/api/pyproject.toml
[project]
name = "footprints-api"
version = "0.1.0"
description = "Footprints FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic-settings>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.39.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "httpx>=0.28.0",
    "ruff>=0.12.0",
    "mypy>=1.17.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
